Update conditions in which libretro-mame will link asmjit.
In `scripts/src/3rdparty.lua`, the library is only defined for the conditions:

```
if (not _OPTIONS["FORCE_DRC_C_BACKEND"]) and ((_OPTIONS["PLATFORM"] == "x86") or (_OPTIONS["PLATFORM"] == "arm64")) then
project "asmjit"
```

However, in `scripts/src/main.lua`, the check for linking it is only:

```
if not _OPTIONS["FORCE_DRC_C_BACKEND"] then
	links {
		"asmjit",
	}
end
```

Which means that for platforms other than `x86` or `arm64`, `-lasmjit` will be
added even though it is not available (unless `FORCE_DRC_C_BACKEND`). This
affects, for instance, compilation for the RISC-V `jh7110` target.

Until this is fixed upstream, this commit provides a patch.